# AutoMM: Avoid redundant checkpointing and evaluation (2x training speedup)